JavaDoc y mejora

+++ src/main/java/com/example/reto1addihibernate/domain/productos/Producto.java
@@ -5,27 +5,48 @@ import jakarta.persistence.*;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
  * Clase que representa un producto en la aplicación.
+ *
+ * Esta clase modela las características de un producto, incluyendo su identificador,
+ * nombre, precio y cantidad disponible en inventario.
  */
 @Data
 @Entity
-@Table(name="producto")
+@Table(name = "producto")
 public class Producto implements Serializable {
+    /**
+     * Identificador único del producto.
+     */
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Column(name="nombre")
+
+    /**
+     * Nombre del producto.
+     */
+    @Column(name = "nombre")
     private String nombre;
-    @Column(name="precio")
-    private int precio;
-    @Column(name="cantidad")
+
+    /**
+     * Precio del producto.
+     */
+    @Column(name = "precio")
+    private double precio;
+
+    /**
+     * Cantidad disponible en inventario del producto.
+     */
+    @Column(name = "cantidad")
     private Integer cantidad;
 
-
+    /**
+     * Representación de cadena del objeto Producto.
+     *
+     * @return Cadena que representa el objeto Producto.
+     */
     @Override
     public String toString() {
         return "Producto{" +
@@ -36,3 +57,4 @@ public class Producto implements Serializable {
                 '}';
     }
 }
+
